Refactor Revit test base classes.
- Merge SystemTest class with RevitSystemTestBase.
- Repoint all SystemTest child classes to RevitSystemTestBase.
- Set DocumentManager UIApplication and UIDocument during setup.
