+++ test/Libraries/RevitIntegrationTests/AdaptiveComponentTests.cs
@@ -5,6 +5,8 @@ using Dynamo.Tests;
 
 using NUnit.Framework;
 
+using RevitTestServices;
+
 using RTF.Framework;
 
 using DoubleSlider = DSCoreNodesUI.Input.DoubleSlider;
@@ -12,7 +14,7 @@ using DoubleSlider = DSCoreNodesUI.Input.DoubleSlider;
 namespace RevitSystemTests
 {
     [TestFixture]
-    class AdaptiveComponentTests : SystemTest
+    class AdaptiveComponentTests : RevitSystemTestBase
     {
         [Test, Category("Failure")]
         [TestModel(@".\AdaptiveComponent\AdaptiveComponentByFace.rfa")]
