+++ test/Libraries/RevitIntegrationTests/SiteLocationTests.cs
@@ -10,6 +10,8 @@ using Revit.GeometryConversion;
 
 using RevitServices.Persistence;
 
+using RevitTestServices;
+
 using RTF.Framework;
 
 using Location = DynamoUnits.Location;
@@ -18,7 +20,7 @@ using SiteLocation = DSRevitNodesUI.SiteLocation;
 namespace RevitSystemTests
 {
     [TestFixture]
-    class SiteLocationTests : SystemTest
+    class SiteLocationTests : RevitSystemTestBase
     {
         [Test, TestModel(@".\SiteLocation\SiteLocation.rvt")]
         public void SiteLocation_ValidArgs()
