+++ test/Libraries/RevitIntegrationTests/XYZTests.cs
@@ -2,12 +2,14 @@
 
 using NUnit.Framework;
 
+using RevitTestServices;
+
 using RTF.Framework;
 
 namespace RevitSystemTests
 {
     [TestFixture]
-    public class XYZTests : SystemTest
+    public class XYZTests : RevitSystemTestBase
     {
         [Test]
         [TestModel(@".\empty.rfa")]
